Project Scholar
This is a 3 page website for my wife's tutoring service. I used  a link navigation to help users travel to each page along with a back to home button at the bottom of each page. On the main page there are link in the navigation that will help you to find specific content on the main Project Scholar page along with an additional link that will take you to a resume page. I also created a form with that showcases 5 different ways to collect information. I used the following types for my form: 
*Text 
*Email
*Phone
*Radio 
*Checkbox
I created additional information box that consist of 4 rows and 50 columns that will allow users to stretch the text box for their convenience.
In the footer I created 3 social media links that are not active. I created 1 media queries that permits adjustments for mobile devices.